Assignment 2/requirements.txt
bs4
re 
requests
chromadb
transformers
mysql-connector-python
langchain_google_genai
sentence_transformers